+++ ontimize/i18n/i18n.ts
@@ -81,7 +81,6 @@ export var MAP = {
 
     'MESSAGES.FORM_VALIDATION_ERROR': 'Form values are not valid!',
     'MESSAGES.FORM_NOTHING_TO_UPDATE_INFO': 'There are no values to update!',
-    'MESSAGES.FORM_CHANGES_WILL_BE_LOST': 'Are you sure you want to continue? All unsaved changes will be lost',
     'MESSAGES.ERROR_LOGIN': 'Login failed. Check username and/or password.',
     'MESSAGES.ERROR_SESSION': 'Session expired.',
     'MESSAGES.ERROR_INSERT': 'Failed to insert data.',
@@ -112,6 +111,8 @@ export var MAP = {
     'TABLE.PAGINATE.LAST': 'Last',
     'TABLE.PAGINATE.NEXT': 'Next',
     'TABLE.PAGINATE.PREVIOUS': 'Previous',
+    'TABLE.PAGINATE.RANGE_LABEL': 'of',
+    'TABLE.PAGINATE.ITEMSPERPAGELABEL': 'Items per page',
     'TABLE.ARIA.SORT_ASCENDING': ': activate to sort column ascending',
     'TABLE.ARIA.SORT_DESCENDING': ': activate to sort column descending',
     'TABLE.BUTTONS.COLVIS': 'Show / hide columns',
@@ -216,9 +217,11 @@ export var MAP = {
     'FORM_VALIDATION.DATE_MIN':'Fecha mínima',
     'FORM_VALIDATION.DATE_MAX':'Fecha máxima',
 
+    'FORM_VALIDATION.FILE_TYPE':'Formato no válido',
+    'FORM_VALIDATION.FILE_MAXSIZE':'Tamaño de fichero máximo',
+
     'MESSAGES.FORM_VALIDATION_ERROR': 'El formulario contiene campos con valores no válidos',
     'MESSAGES.FORM_NOTHING_TO_UPDATE_INFO': '¡No existen valores para actualizar!',
-    'MESSAGES.FORM_CHANGES_WILL_BE_LOST': '¿Está seguro de que quiere continuar? Perderá los cambios realizados',
     'MESSAGES.ERROR_LOGIN': 'Error de inicio de sesión. Compruebe nombre de usuario y/o contraseña.',
     'MESSAGES.ERROR_SESSION': 'La sesión ha caducado',
     'MESSAGES.ERROR_INSERT': 'Error al insertar datos.',
@@ -249,6 +252,8 @@ export var MAP = {
     'TABLE.PAGINATE.LAST': 'Último',
     'TABLE.PAGINATE.NEXT': 'Siguiente',
     'TABLE.PAGINATE.PREVIOUS': 'Anterior',
+    'TABLE.PAGINATE.RANGE_LABEL': 'de',
+    'TABLE.PAGINATE.ITEMSPERPAGELABEL': 'Registros por página',
     'TABLE.ARIA.SORT_ASCENDING': ': activar para ordenar la columna de forma ascendente',
     'TABLE.ARIA.SORT_DESCENDING': ': activar para ordenar la columna de forma descendente',
     'TABLE.BUTTONS.COLVIS': 'Mostrar / ocultar columnas',